requirements.txt
idna==3.3
Django==4.2.7
python-dotenv==0.20.0
cryptography==41.0.7
pyjwt==2.8.0
python-jose==3.3.0
django-oauth-toolkit==2.3.0
django-cors-headers==3.11.0
django-cors-middleware==1.5.0
whitenoise==6.0.0
dj-database-url==1.0.0
psycopg2-binary==2.9.3
django-allauth==0.57.0
gunicorn==21.2.0
tldextract==5.0.1
djangorestframework==3.14.0